Code restructure and style additions to the journey feature
