+++ src/controllers/index.ts
@@ -26,6 +26,8 @@ export const Controllers = createRegistry({
   IconSelector: () => import("@/controllers/icon-selector").then(m => m.IconSelector),
   HierarchyTree: () => import("@/controllers/hierarchy-tree").then(m => m.HierarchyTree),
   IceEditor: () => import("@/controllers/ice-editor").then(m => m.IceEditor),
+  JourneyEditor: () => import("@/controllers/journey-editor").then(m => m.JourneyEditor),
+  JourneysOverview: () => import("@/controllers/journeys-overview").then(m => m.JourneysOverview),
   LabelsEditor: () => import("@/controllers/labels-editor").then(m => m.LabelsEditor),
   LabelCreator: () => import("@/controllers/label-creator").then(m => m.LabelCreator),
   LakesEditor: () => import("@/controllers/lakes-editor").then(m => m.LakesEditor),
@@ -57,15 +59,13 @@ export const Controllers = createRegistry({
   RouteEditor: () => import("@/controllers/route-editor").then(m => m.RouteEditor),
   RouteGroupsEditor: () => import("@/controllers/route-groups-editor").then(m => m.RouteGroupsEditor),
   RoutesOverview: () => import("@/controllers/routes-overview").then(m => m.RoutesOverview),
-  JourneysOverview: () => import("@/controllers/journeys-overview").then(m => m.JourneysOverview),
-  JourneyEditor: () => import("@/controllers/journey-editor").then(m => m.JourneyEditor),
-  TransportTypesEditor: () => import("@/controllers/transport-types-editor").then(m => m.TransportTypesEditor),
   StatesEditor: () => import("@/controllers/states-editor").then(m => m.StatesEditor),
   SubmapTool: () => import("@/controllers/submap-tool").then(m => m.SubmapTool),
   TemperatureGraph: () => import("@/controllers/temperature-graph").then(m => m.TemperatureGraph),
   TradeAnimationEditor: () => import("@/controllers/trade-animation-editor").then(m => m.TradeAnimationEditor),
   TradeDetails: () => import("@/controllers/trade-details").then(m => m.TradeDetails),
   TransformTool: () => import("@/controllers/transform-tool").then(m => m.TransformTool),
+  TransportTypesEditor: () => import("@/controllers/transport-types-editor").then(m => m.TransportTypesEditor),
   UnitsEditor: () => import("@/controllers/units-editor").then(m => m.UnitsEditor),
   View3d: () => import("@/controllers/view-3d").then(m => m.View3d),
   WorldConfigurator: () => import("@/controllers/world-configurator").then(m => m.WorldConfigurator),
