+++ src/index.html
@@ -767,6 +767,7 @@
             <option value="gridOverlay">Grid</option>
             <option value="terrs">Heightmap</option>
             <option value="ice">Ice</option>
+            <option value="journeys">Journeys</option>
             <option value="labels">Labels</option>
             <option value="lakes">Lakes</option>
             <option value="landmass">Landmass</option>
@@ -2907,6 +2908,18 @@
         </marker>
         <marker id="end-arrow-small" viewBox="0 -5 10 10" refX="6" markerWidth="2" markerHeight="2" orient="auto">
           <path d="M0,-5L10,0L0,5" fill="#555" />
+        </marker>
+        <marker
+          id="journey-arrow"
+          viewBox="0 0 10 10"
+          refX="8"
+          refY="5"
+          markerWidth="5"
+          markerHeight="5"
+          markerUnits="strokeWidth"
+          orient="auto-start-reverse"
+        >
+          <path d="M 0 0 L 10 5 L 0 10 z" fill="context-stroke" />
         </marker>
         <g id="defs-relief">
           <symbol id="relief-mount-1" viewBox="0 0 100 100">
